refactor(attack-log): memoize attack definitions base groups
